+++ bindings/matrix-sdk-crypto-ffi/src/machine.rs
@@ -1464,6 +1464,7 @@ impl OlmMachine {
             progress_listener.on_progress(progress as i32, total as i32)
         };
 
+        #[allow(deprecated)]
         let result =
             self.runtime.block_on(self.inner.import_room_keys(keys, from_backup, listener))?;